feat: add RecommendationShelf type and shuffleArray utility; implement HeroDisplay component
