+++ frontend/src/types.d.ts
@@ -1,0 +1,8 @@
+declare namespace PerPlexed {
+    interface RecommendationShelf {
+        title: string;
+        libraryID: string;
+        dir: string;
+        link: string;
+    }
+}
